Remove patch for inaccessible disks; add remark in README file
The original patch for better support for inaccessible disks based
on problems with usb sticks. But this option 'inaccessible' is only
usable, if hard disks can not be accessed by the BIOS, but always
by linux. This option is not sensible for usb sticks, which will
always bring failures if they are named in the lilo.conf.